+++ frontend/src/stores/mswModeStore.ts
